requirements.txt
Django >= 1.11
django-cms >= 3.4
djangocms-text-ckeditor